[Refactor]변수 명 상황에 맞게 수정
