+++ src/main/java/io/mohajistudio/tangerine/prototype/domain/place/controller/RepresentService.java
@@ -4,6 +4,6 @@ import io.mohajistudio.tangerine.prototype.infra.place.dto.AddressDTO;
 
 import java.util.List;
 
-public interface RepresentativePlaceGeneratorInterface {
-    List<String> generate(List<AddressDTO> places);
+public interface RepresentService {
+    List<String> extract(List<AddressDTO> places);
 }
