+++ src/main/java/io/mohajistudio/tangerine/prototype/domain/place/dto/RepresentativeRegionDTO.java
@@ -3,7 +3,6 @@ package io.mohajistudio.tangerine.prototype.domain.place.dto;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 
 import java.util.List;
@@ -11,7 +10,7 @@ import java.util.List;
 @Getter
 @Setter
 @NoArgsConstructor
-public class Region {
+public class RepresentativeRegionDTO {
     @NotNull
     private List<String> regions;
 }
